chore(api): move sentry to its own settigs file
